rename NodeStats to NodeDetails
